Tests/LibWeb/Text/expected/Worker/WorkerGlobalScope-inherited-properties.txt

self.importScripts is own property: false
idlharness would call null-this importScripts: false
done
